docker/api/requirements.txt
gunicorn 
django == 3.2
psycopg2-binary
djangorestframework
